Replace text-based progress bar

+++ DMR11.Core/Base/ITitle.cs
@@ -52,6 +52,6 @@ namespace DMR11.Core
             set;
         }
         
-        Task PopulateChapterAsync(Progress<int> progress);
+        Task PopulateChapterAsync(Progress<double> progress);
     }
 }
